+++ backend/.env
@@ -1,5 +1,5 @@
 PORT = 3001
 DATABASE = bdbikepass
 BDUSERNAME = postgres
-PASSWORD = 270704
+PASSWORD = 123
 FOLDERPHOTOS = ../../public/photos
