--- frontend/src/pages/home.css
@@ -1,51 +1,0 @@
-h1, h2, h3 {
-    margin: 0;
-}
-
-header {
-    display: flex;
-    justify-content: space-between;
-}
-
-main{
-    background-color: cornsilk;
-}
-
-.container{
-    display: flex;
-    justify-content: space-between;
-}
-
-.card{
-    text-align: center;
-    background-color:bisque;
-    border-radius: 15px;
-    height: 398px;
-    width: 248px;
-    padding: 26px;
-    margin: 20px;
-    max-height: 450px;
-}
-
-.cardcontent {
-    background-color:cornsilk;
-    text-align: left;
-}
-
-.cardphoto{
-    background-color: #D9D9D9;
-    height: 180px;
-    max-height: 180px;
-    width: 248px;
-    max-width: 248px;
-}
-
-.descritivo{
-    background-color: #D9D9D9;
-    height: 70px;
-    max-height: 70px;
-    width: 228px;
-    max-width: 228px;
-    border-radius: 15px;
-    padding: 10px;
-}
