+++ frontend/src/routes/demo/Demo.css
@@ -1,4 +1,6 @@
 *{
+    margin: 0;
+    padding: 0;
     font-family: 'Times New Roman', Times, serif;
 }
