+++ frontend/src/routes/pages/modelo.css
@@ -1,3 +1,9 @@
+*{
+  margin: 0;
+  padding: 0;
+  font-family: 'Times New Roman', Times, serif;
+}
+
 .conteudo {
   text-align: center;
 }
